Spring clean and restructuring
Deleted unused code

Made subfolders in the componenet folder, to better navigate the folder
