+++ lib/view/home/home_web.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/cupertino.dart';
-import 'package:temp/components/background_structure.dart';
+import 'package:temp/components/structure/background_structure.dart';
 
 import 'home_content.dart';
